feat: show info snackbar on app start if drone radar is active
DT-2256
